requirements.txt
django==3.0
djangorestframework==3.11.0
gunicorn==19.9.0
